+++ whois-api/src/main/java/net/ripe/db/whois/api/httpserver/dos/WhoisDoSFilter.java
@@ -11,6 +11,7 @@ import org.eclipse.jetty.servlets.DoSFilter;
 import org.eclipse.jetty.util.StringUtil;
 import org.eclipse.jetty.util.annotation.Name;
 import org.slf4j.Logger;
+import net.ripe.db.whois.common.ip.IpInterval;
 
 import java.io.IOException;
 import java.util.Set;
@@ -45,7 +46,27 @@ public abstract class WhoisDoSFilter extends DoSFilter {
 
     @Override
     protected boolean checkWhitelist(final String candidate) {
-        return IpUtil.isExistingIp(candidate, ipv4whitelist, ipv6whitelist);
+        final IpInterval<?> parsed = IpInterval.parse(candidate);
+        return switch (parsed) {
+            case Ipv4Resource ipv4Resource -> {
+                for (Ipv4Resource entry : ipv4whitelist) {
+                    if (entry.contains(ipv4Resource)) {
+                        yield true;
+                    }
+                }
+                yield false;
+            }
+            case Ipv6Resource ipv6Resource -> {
+                for (Ipv6Resource entry : ipv6whitelist) {
+                    if (entry.contains(ipv6Resource)) {
+                        yield true;
+                    }
+                }
+                yield false;
+            }
+            default -> false;
+        };
+
     }
 
     @Override
